Splitted the previous DataFlow class into two: Pump and Flow. The Flow contains the sync related code with the data queue, and the Pump contains the async related code with the callback queue. So updating the data queue won't have any effect on the callback queue until the pump is not refreshed.  I renamed almost every method on the Flow. I use extract instead of read, sustain instead of write, dry instead of empty, and blocked instead of exhausted. I added a new drain method to make the flow dry with a sync loop. #107
I added multiple arguments to the sustain method. These arguments are pushed one by one to the data queue, so one argument means one extract. If you need your data to travel in a single package, then you need to use an array or a data transfer object. #103

With the new system it is possible to read and write multiple data chunks at once. Currently it is not possible to get info about how many messages the flow queue contains, that can be a new feature. #101

I added new examples and the readme is overridden as well. It's time to go to 0.8.1.